Update searchy.scss
I overhauled the look of Searchy. The light blue color is the same color as the header of the google form, but it's all using the SCSS variables so you can change it to whatever in one step (just change the color in Google Forms to match!). I kinda like it, though.

I made these changes totally within Developer Tools in Firefox, and copied them over to my copy of the main stylesheet over here... so it's possible I may have missed something. Here's what it's supposed to look like:
https://imgur.com/a/03p5Cji

So, if it doesn't look like that, then I must have neglected to include one or more lines. I'll leave the tab open overnight so that I am able to dig through my temporary changes and hunt down missing code should that end up being the case.